Added convertOptions and refactored core to remove duplicated convert code
