--- src/types/toBase64Response.ts
@@ -1,5 +1,0 @@
-export type ToBase64Response = {
-  base64?: string;
-  size?: string;
-  page?: number;
-}
