--- src/types/writeImageResponse.ts
@@ -1,7 +1,0 @@
-export type WriteImageResponse = {
-  name?: string;
-  size?: string;
-  fileSize?: number;
-  path?: string;
-  page?: number;
-}
